refactor - Moved all the interfaces to their respective packages.
