+++ charts/src/main/java/dev/anirban/charts/circular/charts/CircularDonutChartColumn.kt
@@ -17,10 +17,10 @@ import dev.anirban.charts.circular.center.CircularNoCenter
 import dev.anirban.charts.circular.colorconvention.CircularGridColorConvention
 import dev.anirban.charts.circular.decoration.CircularDecoration
 import dev.anirban.charts.circular.foreground.CircularDonutForeground
-import dev.anirban.charts.circular.interfaces.CircularCenterInterface
-import dev.anirban.charts.circular.interfaces.CircularColorConventionInterface
-import dev.anirban.charts.circular.interfaces.CircularDataInterface
-import dev.anirban.charts.circular.interfaces.CircularForegroundInterface
+import dev.anirban.charts.circular.center.CircularCenterInterface
+import dev.anirban.charts.circular.colorconvention.CircularColorConventionInterface
+import dev.anirban.charts.circular.data.CircularDataInterface
+import dev.anirban.charts.circular.foreground.CircularForegroundInterface
 
 /**
  * This class is the sub - class of [CircularChart] class which is the root parent class of the
