+++ charts/src/main/java/dev/anirban/charts/circular/data/CircularDataInterface.kt
@@ -1,13 +1,12 @@
-package dev.anirban.charts.circular.interfaces
+package dev.anirban.charts.circular.data
 
-import dev.anirban.charts.circular.data.*
 
 /**
  * This implementation needs to be implemented by all the classes which provide an implementation to
  * handle data
  *
  * Implementations for this interface is :- [CircularDonutListData],[CircularDonutTargetData],
- * [CircularRingTargetData],[CircularTargetDataBuilder]
+ * [CircularTargetDataBuilder]
  *
  * @property itemsList This is the list of items which are shown in the readings
  * @property sweepAngles This is the list of sweep angles which would be used to draw the readings
