+++ charts/src/main/java/dev/anirban/charts/circular/exceptions/CircularDecorationMismatch.kt
@@ -1,7 +1,7 @@
 package dev.anirban.charts.circular.exceptions
 
 import dev.anirban.charts.circular.decoration.CircularDecoration
-import dev.anirban.charts.circular.interfaces.CircularDataInterface
+import dev.anirban.charts.circular.data.CircularDataInterface
 
 /**
  * This class is used to throw an exception when the [CircularDecoration.colorList] size is lesser
